# docs/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Adwais.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/monitors/analytics": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Unified analytics endpoint for monitoring data.\r\nProvides latency time-series and monitoring KPIs for the specified timeframe (defaults to T30).",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific tenant. \r\nIf null, metrics represent the global portfolio total.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "monitorId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific monitor. \r\nIf null, metrics represent the global portfolio total.\r\nOverrides tenantId parameter.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "The primary timeframe for the calculation (e.g., T7, T30). \r\nDefaults to T30.",
            "schema": {
              "$ref": "#/components/schemas/Timeframe"
            }
          },
          {
            "name": "comparison",
            "in": "query",
            "description": "The comparison period for the timeframe.\r\nDefaults to Preceding.",
            "schema": {
              "$ref": "#/components/schemas/ComparisonType"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Optional. Filters the monitors by specific tags.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "excludedTags",
            "in": "query",
            "description": "Optional. Excludes monitors matching any of these tags. Exclusion takes precedence.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "description": "Optional. Filters the monitors by specific statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "excludedStatuses",
            "in": "query",
            "description": "Optional. Excludes monitors with any of these statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAnalyticsResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAnalyticsResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAnalyticsResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/availability": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Returns daily availability for the selected fleet, tenant, or monitor scope.",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific tenant. \r\nIf null, metrics represent the global portfolio total.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "monitorId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific monitor. \r\nIf null, metrics represent the global portfolio total.\r\nOverrides tenantId parameter.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "The primary timeframe for the calculation (e.g., T7, T30). \r\nDefaults to T30.",
            "schema": {
              "$ref": "#/components/schemas/Timeframe"
            }
          },
          {
            "name": "comparison",
            "in": "query",
            "description": "The comparison period for the timeframe.\r\nDefaults to Preceding.",
            "schema": {
              "$ref": "#/components/schemas/ComparisonType"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Optional. Filters the monitors by specific tags.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "excludedTags",
            "in": "query",
            "description": "Optional. Excludes monitors matching any of these tags. Exclusion takes precedence.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "description": "Optional. Filters the monitors by specific statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "excludedStatuses",
            "in": "query",
            "description": "Optional. Excludes monitors with any of these statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAvailabilitySeriesResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAvailabilitySeriesResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAvailabilitySeriesResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/monitors/unassigned": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Retrieves monitors that are not assigned to any specific tenant.",
        "parameters": [
          {
            "name": "timeframe",
            "in": "query",
            "description": "The timeframe for calculating uptime percentage.",
            "schema": {
              "$ref": "#/components/schemas/Timeframe"
            }
          },
          {
            "name": "comparison",
            "in": "query",
            "description": "The comparison period type.",
            "schema": {
              "$ref": "#/components/schemas/ComparisonType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/{id}/assign/{tenantId}": {
      "patch": {
        "tags": [
          "Monitor"
        ],
        "summary": "Reassigns a monitor to a different tenant.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/{id}/unassign": {
      "patch": {
        "tags": [
          "Monitor"
        ],
        "summary": "Moves a monitor to the unassigned (system) tenant.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparisonType": {
        "enum": [
          "Preceding",
          "YearOverYear"
        ],
        "type": "string"
      },
...
      "MonitorAnalyticsResponseDto": {
        "required": [
          "globalAverageLatency",
          "kpis",
          "latencyPoints"
        ],
        "type": "object",
        "properties": {
          "globalAverageLatency": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latencyPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LatencyPointResponseDto"
            },
            "nullable": true
          },
          "kpis": {
            "$ref": "#/components/schemas/MonitorKpiResponseDto"
          }
        },
        "additionalProperties": false
      },
...
      "MonitorAvailabilitySeriesResponseDto": {
        "required": [
          "averageUptimePercentage",
          "lowestUptimePercentage",
          "periodEnd",
          "periodStart",
          "points"
        ],
        "type": "object",
        "properties": {
          "periodStart": {
            "type": "string",
            "format": "date-time"
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "averageUptimePercentage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lowestUptimePercentage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitorAvailabilityPointResponseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "Timeframe": {
        "enum": [
          "Today",
          "T7",
          "T30",
          "T90",
          "Ytd",
          "T365"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your valid token in the text input below.\r\n\r\nExample: \"Bearer eyJhbGciOiJIUzI1Ni...\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}